some minor refactoring to cleanup dependencies
